+++ NEWS.md
@@ -1,3 +1,7 @@
+# queries 0.2.0
+- Improve logic for looking for files
+- Add default parameters to parameters
+
 # queries 0.1.1
 * Changed look for queries in /sql first
 * Parameters have descriptions now